src/requirements.txt
fastapi
fastapi-utilities
uvicorn
pytz
requests
psycopg2-binary